move state sync logic from useAuth hook to AuthSync component
